feat: Moving admin check in an helper
