+++ packages/front/pages/admin.vue
@@ -18,16 +18,8 @@
 
 <script setup lang="ts">
 import { onBeforeMount } from "vue";
-import { useUserStore } from "@/stores/user";
-import { usePopUpsStore } from "@/stores/popups";
-import { useRouter } from "vue-router";
+import { adminCheck } from "~/helpers/adminCheck";
 import { testApi } from "~/api/test.api";
-
-import type { UserDatas } from "@/types/User.d";
-
-const userStore = useUserStore();
-const popUpsStore = usePopUpsStore();
-const router = useRouter();
 
 const isLoading = ref(false);
 
@@ -40,20 +32,6 @@ const checkToken = async () => {
 };
 
 onBeforeMount(() => {
-  let isAdmin: boolean = false;
-
-  if (typeof window !== "undefined") {
-    const userDatas: { datas: UserDatas; isAdmin: boolean } = JSON.parse(
-      localStorage.getItem("userData") ?? "{}"
-    );
-    isAdmin = userDatas.isAdmin;
-  } else {
-    isAdmin = userStore.isAdmin;
-  }
-
-  if (!isAdmin) {
-    popUpsStore.setPopUp(new Error("Vous n'êtes pas admin!"), "error");
-    router.push("/login");
-  }
+  adminCheck();
 });
 </script>
